feat: the implementation of the chat creation has moved to usecase
