+++ crates/adapters/src/api/chat/chat_controllers.rs
@@ -3,8 +3,6 @@ use uuid::Uuid;
 
 use application::usecases::chat::create::CreateChatUseCase;
 use application::usecases::interfaces::AbstractUseCase;
-use application::usecases::user_chat::create::CreateUserChatUseCase;
-use domain::user_chat_entity::UserChatEntity;
 
 use crate::api::shared::app_state::AppState;
 use crate::api::shared::error_presenter::ErrorReponse;
@@ -16,25 +14,11 @@ pub fn routes(cfg: &mut web::ServiceConfig) {
 #[post("/chat/create")]
 //todo receive user id by token
 async fn create_chat(data: web::Data<AppState>, info: web::Json<Uuid>) -> Result<HttpResponse, ErrorReponse> {
-    let create_chat_usecase = CreateChatUseCase::new(&data.chat_repository);
+    let create_chat_usecase =
+        CreateChatUseCase::new(info.0, &data.chat_repository, &data.user_chat_repository);
+    let user = create_chat_usecase.execute().await;
 
-    let chat = create_chat_usecase.execute().await;
-
-    match chat {
-        Ok(chat) => {
-            let user_chat = UserChatEntity::new(info.0, chat.id);
-
-            let create_user_chat_usecase =
-                CreateUserChatUseCase::new(user_chat, &data.user_chat_repository);
-
-            let result = create_user_chat_usecase.execute().await;
-
-            result
-                .map_err(ErrorReponse::map_io_error)
-                .map(|_| HttpResponse::Ok().finish())
-        }
-        Err(api_error) => {
-            Err(ErrorReponse::map_io_error(api_error))
-        }
-    }
+    user
+        .map_err(ErrorReponse::map_io_error)
+        .map(|_| HttpResponse::Ok().finish())
 }
