+++ crates/application/src/repositories/user_repository_abstract.rs
@@ -1,7 +1,8 @@
 use std::error::Error;
 
-use domain::user_entity::UserEntity;
 use uuid::Uuid;
+
+use domain::user_entity::UserEntity;
 
 pub trait UserRepositoryAbstract {
     fn save(&self, user: &UserEntity) -> Result<(), Box<dyn Error>>;
