+++ crates/application/src/usecases/mod.rs
@@ -1,4 +1,3 @@
 pub mod user;
 pub mod chat;
-pub mod user_chat;
 pub mod interfaces;
