--- crates/application/src/usecases/user_chat/mod.rs
@@ -1,1 +1,0 @@
-pub mod create;
