+++ crates/db/Cargo.toml
@@ -10,4 +10,4 @@ diesel = { version = "2.1.4", features = ["postgres", "r2d2", "uuid", "chrono"] 
 diesel_migrations = "2.1.0"
 dotenv = "0.15"
 uuid = { version = "1.6.1", features = ["v4"] }
-chrono = {version = "0.4.31", features = ["unstable-locales"]}
+chrono = { version = "0.4.31", features = ["unstable-locales"] }
