+++ crates/domain/Cargo.toml
@@ -6,4 +6,4 @@ publish = false
 
 [dependencies]
 uuid = { version = "1.6.1", features = ["v4"] }
-chrono = {version = "0.4.31", features = ["unstable-locales"]}
+chrono = { version = "0.4.31", features = ["unstable-locales"] }
