+++ crates/domain/src/user_entity.rs
@@ -1,6 +1,5 @@
 use uuid::Uuid;
 
-
 #[derive(Debug, Clone)]
 pub struct UserEntity {
     pub id: Uuid,
